backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.24",
    "python-socketio>=5.16.1",
    "redis>=7.4.0",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.44.0",
]
